Remove min from buckets propType
Warning: As of Prebid.js 3.0, the ‘min’ parameter is no longer supported in custom granularities.

Also note an important idiosyncracy of the way that price ranges are supported: the interval starts at the min value, and the max of one range defines the min of the next range. So if the second interval defines an implicit min of 0.99 and goes to 5 with an increment of 0.05, Prebid.js will generate the values: 0.99, 1.04, 1.09, etc.

This implies that ranges should have max values that are really the min value of next range.